Refactor how dataload processes are defined
This also renames Process to ProcessRun, and uses process to refer to
e.g. the general process of converting a file, rather than a particular
conversion (which is now a run).

Processes for resourceprojects are now defined in a single OrderedDict
(PROCESSES) at the top of views.py.
The plan is to move this resourceprojects specific code out to a
separate repository, see
https://github.com/NRGI/resource-projects-etl/issues/40

Some computation is moved from the templates to the view, now that we
can iterate over PROCESSES. The hope is that this will make implementing
correct status information for deletion easier
(https://github.com/OpenDataServices/cove/issues/112 ).