[#46] Add validation error table.
Add table within a panel that shows frequency and location of first 3 validation errors for each error message.

Refactor functional tests slightly so that you can provide a list of strings to be found in page instead of needing for the whole test to be run again for different string.