[#296] Catch stats producing errors.
Catch all main errors caused when decending json objects and do not
show stats if this is the case.  Give warning to user saying this.

Errors will still be raised if the data passes validation but
there is an error in the stats code. So basically we do a good attempt
at producing stats for invalid data but only if the data is close
to being valid.