Allows html in error messages
I wanted to make one of the error messages look nicer- specifically when
tender_releases_2_releases_not_json.json was loaded
To do this I think the error message in views.py needs to contain html
I've added the | safe option to the html template

You may not think this is a good idea! Please review.