[#469] Fix error on required field
The code did not take into account where there is a sub object
with a required field, assumed only it would be a list of objects.
Due to this path items where ints and was assuemed to be stings causing
the error.  The code takes this into account now.

Also deletes useless print statement.