[#528] codelist verification.
Looks up codelists from github and caches them.
Adds two new boxes to frontend one for closed and one for open
codelists.
Removes validation results for enums that also have codelists. This was
a bit messy as it was hard to get out correct properties from jsonschema
when dealing with arrays of enums.