requirements: Remove rfc6266
It's not needed any more; it is added by
https://github.com/OpenDataServices/cove/commit/4922fddec5f516cdf3e1109b3fa3ddb8edca0073
The functionality was moved to lib-cove-web
And later versions of that package used other libraries

It needs to be done as it still has use_2to3 in setup, and that causes later version of
python tools to crash. This is stopping us upgrading this app to later versions
of Python / host.